Added serializers for the transaction models
